NP-761 Refactor discipline search
